Create Board.cs
Added a board class that will represent the board during games.
This class is made internal because it is only used by other classes in the Engine.
All methods have a description.
SetCell(x, y, type) is used to set a cell equal to the given type.
EmptyCell(x, y) is used to empty a cell.
CellIsSet(x, y) returns whether the given cell is already set to a type.