Added MakeMove(heading) and Heading enum to TetrisGame
TetrisGame now has a instance method called MakeMove() which can be used to try to make a move. It returns true or false depending on whether the move was successful. Unsuccessful moves can be a result of collision detection.
Heading enum has 3 options, LEFT, DOWN or RIGHT.

MakeMove uses 2 queues to empty and fill cells depending on the direction of the move.